// src/interfaces/patient.ts
export interface PatientInterface {
  account_id: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  phoneNumber: string;
  patId: string;
  pat_id: string;
  fname: string;
  lname: string;
  area_code: string;
  add_date: string;
  phone_id: number;
  phone_no: string;
  email: string;
  birth_date: string;
  addr_id: number;
  addr1: string;
  addr2: string;
  city: string;
  state_cd: string;
  zip: string;
  gender_cd: string;
  country: string;
  phone_nbr: string;
  pt_name: string;
  prescriber: string;
  blue_view: boolean;
  active_yn: number;
  allergies: [];
  cardId: string;
  expiration: string;
  lastFourDigits: string;
  type: string;
  hp_area_code: string;
  hp_phone_number: string;
}